разделит файл на новые модули: get-random: две функции, которые возвращают случайное число и случайный элемент массива create-comments: создает комментарий к фотографии create-descriptions: создает описание к фотографии
