+++ js/get-random.js
@@ -1,0 +1,3 @@
+const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+const getRandomArrayElement = (array) => array[getRandomNumber(0, array.length - 1)];
+export{getRandomNumber, getRandomArrayElement};
